requirements.txt
Flask==2.2.2
joblib==1.2.0
keras==2.10.0
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.0
Pillow==9.2.0
scikit_learn==1.1.2
tensorflow==2.10.0
torch==1.12.1
torchvision==0.13.1
